refactor member annotation import
At the moment we have two ways of importing annotations.
1. For classes there is the `DeclarationHandler` that will register `onDeclaredAnnotations(annotations)` for which at a later point in time when we have all the information we will call `JavaClass.completeAnnotations(context)`.
2. For methods on the other hand we will create a memoizing `Supplier` that will finish building the annotations from the context on the first call of `getAnnotations()`.
So far this has usually be fine, because the getter will only be called later when the full class graph has been constructed and all the information is present. However, this is quite hard to understand, even more so as it is inconsistent with the way we do it for `JavaClass`. Also if the getter at some point is called to early during the import process, because some logic changed there, we will end up with a possibly corrupt state, since the memoizing `Supplier` will remember the state at the first invocation.
I feel this is not very understandable for future maintenance. Also it makes it hard to know when exactly a "legal" time has come to call the getter. Thus I have refactored it to use exactly the same pattern as for `JavaClass`. I.e. at a certain step in the import process we will complete all the annotations of all `JavaClasses`, but we will now *also* do this for all members.
This will also make it easier in the following steps to understand when the annotation construction is taking place and when the last possible point in time to recursively import all meta-annotations has come.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>