squash me: Import of default values of annotation methods
These changes guarantee now that default values
in annotations are imported, too. The problem was that when the default
values were to be completed, their respective methods have not been
completed yet (But were already visited). Thus, the loop over the
methods was empty. I dediced to put the method completion just before
that but I do not know if that is the best way to do this. Furthermore,
I extended the tests, so default values are checked, too.

Signed-off-by: Korbinian Singhammer <korbinian.singhammer@tngtech.com>